Move SelectedWallet component
